+++ framework/modules/news/controllers/newsController.php
@@ -237,7 +237,7 @@ class newsController extends expController {
    	 *
    	 * @return string
    	 */
-   	function aggregateWhereClause() {
+   	function aggregateWhereClause($type='') {
         $sql = parent::aggregateWhereClause();
         $sql = "(publish = 0 or publish <= " . time() . ") AND (unpublish=0 OR unpublish > ".time().") AND ".$sql;
         if (isset($this->config['only_featured'])) $sql .= ' AND is_featured=1';